[CodeCompletion] Move solver-based DotExpr completion to its own file
